Parameterize addr.rs
