+++ kernel-rs/src/arch/addr.rs
@@ -1,12 +1,31 @@
 use core::ops::Add;
 
-use bitflags::bitflags;
-
-/// Bytes per page.
-pub const PGSIZE: usize = 4096;
-
 /// Bits of offset within a page.
 pub const PGSHIFT: usize = 12;
+
+/// Bytes per page.
+pub const PGSIZE: usize = 1 << PGSHIFT;
+
+/// Bits of offset for each page table level.
+pub const PLSHIFT: usize = 9;
+
+/// Bytes per page table level.
+pub const PLSIZE: usize = 1 << PLSHIFT;
+
+/// Bit mask for page table index.
+pub const PLMASK: usize = PLSIZE - 1;
+
+/// The number of page table levels.
+pub const PLNUM: usize = 3;
+
+/// Bit position of the page number in PTE.
+pub const PTESHIFT: usize = 10;
+
+/// One beyond the highest possible virtual address.
+/// MAXVA is actually one bit less than the max allowed by
+/// Sv39, to avoid having to sign-extend virtual addresses
+/// that have the high bit set.
+pub const MAXVA: usize = 1 << (PLSHIFT * PLNUM + PGSHIFT - 1);
 
 #[inline]
 pub const fn pgroundup(sz: usize) -> usize {
@@ -18,47 +37,16 @@ pub const fn pgrounddown(a: usize) -> usize {
     a & !PGSIZE.wrapping_sub(1)
 }
 
-bitflags! {
-    pub struct PteFlags: usize {
-        /// valid
-        const V = 1 << 0;
-        /// readable
-        const R = 1 << 1;
-        /// writable
-        const W = 1 << 2;
-        /// executable
-        const X = 1 << 3;
-        /// user-accessible
-        const U = 1 << 4;
-    }
-}
-
 /// Shift a physical address to the right place for a PTE.
 #[inline]
 pub fn pa2pte(pa: PAddr) -> usize {
-    (pa.into_usize() >> 12) << 10
+    (pa.into_usize() >> PGSHIFT) << PTESHIFT
 }
 
 #[inline]
 pub fn pte2pa(pte: usize) -> PAddr {
-    ((pte >> 10) << 12).into()
+    ((pte >> PTESHIFT) << PGSHIFT).into()
 }
-
-/// Extract the three 9-bit page table indices from a virtual address.
-
-/// 9 bits
-pub const PXMASK: usize = 0x1ff;
-
-#[inline]
-pub fn pxshift(level: usize) -> usize {
-    PGSHIFT + 9 * level
-}
-
-/// One beyond the highest possible virtual address.
-/// MAXVA is actually one bit less than the max allowed by
-/// Sv39, to avoid having to sign-extend virtual addresses
-/// that have the high bit set.
-pub const MAXVA: usize = (1) << (9 + 9 + 9 + 12 - 1);
 
 pub trait Addr: Copy + From<usize> + Add<usize, Output = Self> {
     fn into_usize(self) -> usize;
@@ -67,7 +55,7 @@ pub trait Addr: Copy + From<usize> + Add<usize, Output = Self> {
 }
 
 pub trait VAddr: Addr {
-    fn px(&self, level: usize) -> usize;
+    fn page_table_index(&self, level: usize) -> usize;
 }
 
 macro_rules! define_addr_type {
@@ -108,8 +96,8 @@ macro_rules! define_addr_type {
 macro_rules! impl_vaddr {
     ($typ:ident) => {
         impl VAddr for $typ {
-            fn px(&self, level: usize) -> usize {
-                (self.into_usize() >> pxshift(level)) & PXMASK
+            fn page_table_index(&self, level: usize) -> usize {
+                (self.into_usize() >> (PGSHIFT + PLSHIFT * level)) & PLMASK
             }
         }
     };
