+++ kernel-rs/src/page.rs
@@ -6,9 +6,15 @@ use core::{
     ptr::NonNull,
 };
 
+use static_assertions::const_assert;
+
 use crate::arch::addr::{PAddr, PGSIZE};
 
+// `RawPage` must be aligned with PGSIZE.
+const_assert!(PGSIZE == 4096);
+
 /// Page type.
+///
 #[repr(align(4096))]
 pub struct RawPage {
     inner: [u8; PGSIZE],
